+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+- **MP4**: Padding atoms (`free`) are used when writing
+
+### Fixed
+- **MP4**: `meta` atoms are written correctly
+
 ## [0.5.0] - 2022-02-20
 
 ### Added